# Improve performance of SemanticTextPartitioner