@stephentoub 2023-04-20
> I wish we had UTs for this

I'm happy to add more unit tests if there's something you'd like to see covered that's not currently covered. Suggestions? (But yes, in the above perf test, I did separately validate that the results were the same before/after.)